+++ packages/headless/src/features/recommendation/recommendation-analytics-actions.ts
@@ -1,4 +1,11 @@
-import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
+import {Result} from '../../api/search/search/result';
+import {
+  AnalyticsType,
+  documentIdentifier,
+  makeAnalyticsAction,
+  partialRecommendationInformation,
+  validateResultPayload,
+} from '../analytics/analytics-utils';
 
 /**
  * Logs a search event with an `actionCause` value of `recommendationInterfaceLoad`.
@@ -8,3 +15,16 @@ export const logRecommendationUpdate = makeAnalyticsAction(
   AnalyticsType.Search,
   (client) => client.logRecommendationInterfaceLoad()
 );
+
+export const logRecommendationOpen = (result: Result) =>
+  makeAnalyticsAction(
+    'analytics/recommendation/open',
+    AnalyticsType.Click,
+    (client, state) => {
+      validateResultPayload(result);
+      return client.logRecommendationOpen(
+        partialRecommendationInformation(result, state),
+        documentIdentifier(result)
+      );
+    }
+  )();