+++ packages/headless/src/recommendation.index.ts
@@ -23,6 +23,7 @@ export * from './features/pipeline/pipeline-actions-loader';
 export * from './features/search-hub/search-hub-actions-loader';
 export * from './features/debug/debug-actions-loader';
 export * from './features/recommendation/recommendation-actions-loader';
+export * from './features/recommendation/recommendation-click-analytics-actions-loader';
 
 // Controllers
 export {
@@ -40,3 +41,5 @@ export {
   ContextPayload,
   buildContext,
 } from './controllers/context/headless-context';
+
+export {Result} from './api/search/search/result';