+++ packages/headless/src/state/state-sections.ts
@@ -25,6 +25,7 @@ import {StateWithHistory} from '../app/undoable';
 import {HistoryState} from '../features/history/history-state';
 import {FoldingState} from '../features/folding/folding-state';
 import {TriggerState} from '../features/triggers/triggers-state';
+import {QuestionAnsweringState} from '../features/question-answering/question-answering-state';
 
 export interface QuerySection {
   /**
@@ -236,4 +237,11 @@ export interface TriggerSection {
    * The information related to the triggers.
    */
   triggers: TriggerState;
+}
+
+export interface QuestionAnsweringSection {
+  /**
+   * The question and answers for a given query, also known as smart snippet.
+   */
+  questionAnswering: QuestionAnsweringState;
 }