+++ packages/headless/src/test/mock-state.ts
@@ -26,6 +26,7 @@ import {getFacetOrderInitialState} from '../features/facets/facet-order/facet-or
 import {getResultPreviewInitialState} from '../features/result-preview/result-preview-state';
 import {getFoldingInitialState} from '../features/folding/folding-state';
 import {getTriggerInitialState} from '../features/triggers/triggers-state';
+import {getQuestionAnsweringInitialState} from '../features/question-answering/question-answering-state';
 
 export function createMockState(
   config: Partial<SearchAppState> = {}
@@ -59,6 +60,7 @@ export function createMockState(
     version: 'unit-testing-version',
     folding: getFoldingInitialState(),
     triggers: getTriggerInitialState(),
+    questionAnswering: getQuestionAnsweringInitialState(),
     ...config,
   };
 }